Move token_to_string to the ast library
