+++ src/ast/CMakeLists.txt
@@ -1,6 +1,12 @@
 add_library(
-    littlec_ast STATIC lexer.cpp location.cpp nodes.cpp parser.cpp
-                       serializer.cpp type.cpp
+    littlec_ast STATIC
+    lexer.cpp
+    location.cpp
+    nodes.cpp
+    parser.cpp
+    serializer.cpp
+    token_to_string.cpp
+    type.cpp
 )
 
 target_compile_options(
