+++ src/ast/token_to_string.hpp
@@ -1,9 +1,9 @@
 #ifndef TOKEN_TO_STRING_HPP
 #define TOKEN_TO_STRING_HPP
 
-#include <string>
+#include "expr_nodes.hpp"
 
-#include <ast/expr_nodes.hpp>
+#include <string>
 
 [[nodiscard]] std::string tok_to_string(ast::binary_expr::operand /*op*/);
 [[nodiscard]] std::string tok_to_string(ast::unary_expr::operand /*op*/);
